feat: update example with typescript
