+++ examples/with-redux-toolkit/package.json
@@ -4,7 +4,8 @@
   "scripts": {
     "dev": "next",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "type-check": "tsc"
   },
   "dependencies": {
     "@reduxjs/toolkit": "1.5.0",
@@ -12,6 +13,14 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-redux": "7.2.2"
+  },
+  "devDependencies": {
+    "@types/node": "14.14.35",
+    "@types/react": "17.0.3",
+    "@types/react-dom": "17.0.2",
+    "@types/react-redux": "7.1.16",
+    "prettier": "2.2.1",
+    "typescript": "4.2.3"
   },
   "license": "MIT"
 }
