--- examples/with-redux-toolkit/src/app/store.js
@@ -1,8 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit'
-import counterReducer from '../features/counter/counterSlice'
-
-export const store = configureStore({
-  reducer: {
-    counter: counterReducer,
-  },
-})
