+++ examples/with-redux-toolkit/src/features/counter/Counter.tsx
