+++ examples/with-redux-toolkit/src/features/counter/counterSlice.ts
@@ -1,6 +1,11 @@
-import { createSlice } from '@reduxjs/toolkit'
+import { createSlice, Dispatch } from '@reduxjs/toolkit'
+import { CoreState } from '../../app/store'
 
-const initialState = {
+type CounterState = {
+  value: number
+}
+
+const initialState: CounterState = {
   value: 0,
 }
 
@@ -31,7 +36,7 @@ export const { increment, decrement, incrementByAmount } = counterSlice.actions
 // can be dispatched like a regular action: `dispatch(incrementAsync(10))`. This
 // will call the thunk with the `dispatch` function as the first argument. Async
 // code can then be executed and other actions can be dispatched
-export const incrementAsync = (amount) => (dispatch) => {
+export const incrementAsync = (amount: number) => (dispatch: Dispatch) => {
   setTimeout(() => {
     dispatch(incrementByAmount(amount))
   }, 1000)
@@ -40,6 +45,6 @@ export const incrementAsync = (amount) => (dispatch) => {
 // The function below is called a selector and allows us to select a value from
 // the state. Selectors can also be defined inline where they're used instead of
 // in the slice file. For example: `useSelector((state: RootState) => state.counter.value)`
-export const selectCount = (state) => state.counter.value
+export const selectCount = (state: CoreState) => state.counter.value
 
 export default counterSlice.reducer
