--- examples/with-redux-toolkit/src/pages/_app.js
@@ -1,13 +1,0 @@
-import { Provider } from 'react-redux'
-import { store } from '../app/store'
-import '../styles/globals.css'
-
-const MyApp = ({ Component, pageProps }) => {
-  return (
-    <Provider store={store}>
-      <Component {...pageProps} />
-    </Provider>
-  )
-}
-
-export default MyApp
