+++ examples/with-redux-toolkit/src/pages/index.tsx
